move HiveEntityFactory class back into metastore package, remove unneccesary Guava exclusion
