+++ circus-train-core/src/test/java/com/hotels/bdp/circustrain/core/metastore/HiveEntityFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hotels.bdp.circustrain.core;
+package com.hotels.bdp.circustrain.core.metastore;
 
 import java.io.File;
 import java.util.ArrayList;
