+++ circus-train-core/src/test/java/com/hotels/bdp/circustrain/core/replica/AddCheckSumReplicaTableFactoryTest.java
@@ -20,10 +20,10 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
 import static com.hotels.bdp.circustrain.api.CircusTrainTableParameter.PARTITION_CHECKSUM;
-import static com.hotels.bdp.circustrain.core.HiveEntityFactory.newFieldSchema;
-import static com.hotels.bdp.circustrain.core.HiveEntityFactory.newPartition;
-import static com.hotels.bdp.circustrain.core.HiveEntityFactory.newStorageDescriptor;
-import static com.hotels.bdp.circustrain.core.HiveEntityFactory.newTable;
+import static com.hotels.bdp.circustrain.core.metastore.HiveEntityFactory.newFieldSchema;
+import static com.hotels.bdp.circustrain.core.metastore.HiveEntityFactory.newPartition;
+import static com.hotels.bdp.circustrain.core.metastore.HiveEntityFactory.newStorageDescriptor;
+import static com.hotels.bdp.circustrain.core.metastore.HiveEntityFactory.newTable;
 import static com.hotels.bdp.circustrain.core.conf.ReplicationMode.FULL;
 
 import java.io.File;
