+++ pom.xml
@@ -166,12 +166,6 @@
         <groupId>com.hotels</groupId>
         <artifactId>hcommon-hive-metastore</artifactId>
         <version>${hcommon-hive-metastore.version}</version>
-        <exclusions>
-          <exclusion>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-          </exclusion>
-        </exclusions>
       </dependency>
 
       <dependency>
